refactor(api): use shared discovery-protocol keys + extract readonly types
CacheReadonlyService now imports REGISTRY_KEY and heartbeatKeyFor()
from @betterdb/shared (added in the parent commit on this branch's
base) instead of redefining the literals locally. Ensures the API
reader and the cache-package writers reference the same constants.

The 12 exported types describing the read-only response shapes
(CacheHealth, CacheListEntry, ThresholdRecommendation, etc.) move
to a sibling cache-readonly.types.ts file. The service file
re-exports them so existing import paths continue to work, but the
types now live separately and the service file shrinks back to
behavior-only.
